Time: 0 ms (100%), Space: 8.4 MB (27.88%) - LeetHub

+++ 0009-palindrome-number/0009-palindrome-number.cpp
@@ -2,11 +2,11 @@ class Solution {
 public:
     bool isPalindrome(int x) {
         if (x < 0) return false;
-        int cpy = x;
+        long long cpy = x;
         long long rev = 0;
 
         while (cpy > 0) {
-            int digit = cpy % 10;
+            long long digit = cpy % 10;
             cpy /= 10;
             rev = rev * 10 + digit;
         }
